+++ tests/test_utils/recipes/h100/unit-tests.yaml
@@ -193,7 +193,14 @@ products:
         scope: [unit-tests]
         n_repeat: [1]
         time_limit: [1800]
-  - test_case: [tests/unit_tests/distributed/megatron_fsdp/**/*.py]
+  - test_case: [tests/unit_tests/distributed/mfsdp_v1/**/*.py]
+    products:
+      - environment: [lts, dev]
+        tag: [latest]
+        scope: [unit-tests]
+        n_repeat: [1]
+        time_limit: [1800]
+  - test_case: [tests/unit_tests/distributed/mfsdp_v2/**/*.py]
     products:
       - environment: [lts, dev]
         tag: [latest]
