+++ tests/unit_tests/distributed/mfsdp_v1/test_mcore_fully_sharded_data_parallel.py
@@ -27,7 +27,7 @@ from megatron.core.tensor_parallel.random import model_parallel_cuda_manual_seed
 from megatron.core.transformer import TransformerConfig
 from megatron.core.transformer.transformer_layer import TransformerLayer
 from megatron.core.utils import is_te_min_version, is_torch_min_version
-from tests.unit_tests.distributed.megatron_fsdp.utils import (
+from tests.unit_tests.distributed.mfsdp_v1.utils import (
     make_gpt_mock_data_iterator,
     make_moe_args_model_and_optimizer,
     pretrain_forward_backward,
